feat: converted the JsonDB shallow option to be depth instead to give users better control of how deep to access nested json data.
refactored out the JsonRecordConsumer from the JsonRecordSupport class since that file was kinda large and easy to get lost in.
