+++ app/rest/jsondb/src/main/java/io/syndesis/jsondb/GetOptions.java
@@ -23,15 +23,15 @@ import io.syndesis.model.ToJson;
 public class GetOptions implements ToJson, Cloneable {
 
     private boolean prettyPrint;
-    private boolean shallow;
+    private Integer depth;
     private String callback;
 
     public boolean prettyPrint() {
         return prettyPrint;
     }
 
-    public boolean shallow() {
-        return shallow;
+    public Integer depth() {
+        return depth;
     }
 
     public String callback() {
@@ -43,8 +43,8 @@ public class GetOptions implements ToJson, Cloneable {
         return this;
     }
 
-    public GetOptions shallow(final boolean shallow) {
-        this.shallow = shallow;
+    public GetOptions depth(final Integer depth) {
+        this.depth = depth;
         return this;
     }
