+++ app/rest/jsondb/src/main/java/io/syndesis/jsondb/impl/JsonRecordSupport.java
@@ -15,32 +15,23 @@
  */
 package io.syndesis.jsondb.impl;
 
+import com.fasterxml.jackson.core.JsonFactory;
+import com.fasterxml.jackson.core.JsonParseException;
+import com.fasterxml.jackson.core.JsonParser;
+import com.fasterxml.jackson.core.JsonToken;
+import io.syndesis.jsondb.GetOptions;
+import io.syndesis.jsondb.JsonDBException;
+
 import java.io.IOException;
 import java.io.InputStream;
 import java.io.OutputStream;
-import java.nio.charset.StandardCharsets;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.HashSet;
-import java.util.LinkedHashSet;
-import java.util.LinkedList;
-import java.util.List;
-import java.util.Set;
 import java.util.function.Consumer;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 import java.util.stream.Collectors;
-
-import com.fasterxml.jackson.core.JsonFactory;
-import com.fasterxml.jackson.core.JsonGenerator;
-import com.fasterxml.jackson.core.JsonParseException;
-import com.fasterxml.jackson.core.JsonParser;
-import com.fasterxml.jackson.core.JsonStreamContext;
-import com.fasterxml.jackson.core.JsonToken;
-import com.fasterxml.jackson.core.JsonTokenId;
-
-import io.syndesis.jsondb.GetOptions;
-import io.syndesis.jsondb.JsonDBException;
 
 import static com.fasterxml.jackson.core.JsonToken.END_ARRAY;
 import static com.fasterxml.jackson.core.JsonToken.END_OBJECT;
@@ -202,7 +193,7 @@ public final class JsonRecordSupport {
         return toLexSortableString(idx, '[');
     }
 
-    private static int toArrayIndex(String value) {
+    protected static int toArrayIndex(String value) {
         return fromLexSortableStringToInt(value, '[');
     }
 
@@ -244,197 +235,4 @@ public final class JsonRecordSupport {
         return rc;
     }
 
-    /* default */ static class JsonRecordConsumer implements Consumer<JsonRecord> {
-
-        private final String base;
-        private final JsonGenerator jg;
-        private final OutputStream output;
-        private final GetOptions options;
-        @SuppressWarnings("JdkObsolete")
-        private final LinkedList<PathPart> currentPath = new LinkedList<>();
-        private final Set<String> shallowObjects = new LinkedHashSet<>();
-
-        /* default */ JsonRecordConsumer(String base, OutputStream output, GetOptions options) throws IOException {
-            this.base = base;
-            this.output = output;
-            try {
-                this.options = options.clone();
-            } catch (CloneNotSupportedException e) {
-                throw new IOException(e);
-            }
-
-            if( this.options.callback()!=null ) {
-                String backack = this.options.callback() + "(";
-                output.write(backack.getBytes(StandardCharsets.UTF_8));
-            }
-            this.jg = new JsonFactory().createGenerator(output);
-            if( options.prettyPrint() ) {
-                jg.useDefaultPrettyPrinter();
-            }
-        }
-
-        @Override
-        public void accept(JsonRecord record) {
-            try {
-                // Handle the end end of stream..
-                if (record == null) {
-                    close();
-                    return;
-                }
-
-                String path = record.getPath();
-                path = Strings.trimSuffix(path.substring(base.length()), "/");
-
-                // Lets see how much of the path we match compared to
-                // when we last got called.
-                List<String> newPath = new ArrayList<>(Arrays.asList(path.split("/")));
-                if (newPath.size() == 1 && newPath.get(0).isEmpty()) {
-                    newPath.clear();
-                }
-
-                // should we skip over deep records?
-                if( this.options.shallow() && newPath.size() > 1 ) {
-                    shallowObjects.add(newPath.get(0));
-                    return;
-                }
-
-                int pathMatches = getPathMatches(newPath);
-
-                // we might need to close objects down...
-                closeDownStructs(pathMatches);
-                // or open some new ones up...
-                openUpStructs(newPath, pathMatches);
-
-                if (!currentPath.isEmpty() && jg.getOutputContext().inArray()) {
-                    PathPart pathPart = currentPath.getLast();
-                    String last = newPath.get(newPath.size() - 1);
-                    int idx = toArrayIndex(last);
-                    while (idx > pathPart.getIdx()) {
-                        // to track the nulls that lead up to the next value.
-                        pathPart.incrementIdx();
-                        jg.writeNull();
-                    }
-                    // to track the value that we are about to write.
-                    pathPart.incrementIdx();
-                }
-
-                writeValue(record);
-
-            } catch (IOException e) {
-                throw new JsonDBException(e);
-            }
-        }
-
-        private void openUpStructs(List<String> newPath, int pathMatches) throws IOException {
-            int count;
-
-            // we might need to open up objects...
-            count = newPath.size();
-            for (int i = pathMatches; i < count; i++) {
-                String part = newPath.get(i);
-                boolean array = part.startsWith("[");
-
-                if (array) {
-                    if (jg.getOutputContext().inRoot()) {
-                        jg.writeStartArray();
-                    }
-                } else {
-                    if (jg.getOutputContext().inRoot()) {
-                        jg.writeStartObject();
-                    }
-                    jg.writeFieldName(part);
-                }
-
-                if (i + 1 < count) {
-                    String nextPart = newPath.get(i + 1);
-                    boolean nextArray = nextPart.startsWith("[");
-
-                    PathPart pathPart = new PathPart(part, nextArray);
-                    currentPath.add(pathPart);
-
-                    if (nextPart.startsWith("[")) {
-                        jg.writeStartArray();
-
-                        int idx = toArrayIndex(nextPart);
-                        while (idx > pathPart.getIdx()) {
-                            pathPart.incrementIdx();
-                            jg.writeNull();
-                        }
-
-                    } else {
-                        jg.writeStartObject();
-                    }
-                }
-            }
-        }
-
-        private void closeDownStructs(int pathMatches) throws IOException {
-            int count = currentPath.size() - pathMatches;
-            for (int i = 0; i < count; i++) {
-                if (currentPath.removeLast().isArray()) {
-                    jg.writeEndArray();
-                } else {
-                    jg.writeEndObject();
-                }
-            }
-        }
-
-        private int getPathMatches(List<String> newPath) {
-            int pathMatches = 0;
-            for (int i = 0; i < currentPath.size() && i < newPath.size(); i++) {
-                PathPart lastPart = currentPath.get(i);
-                if (lastPart.getPath().equals(newPath.get(i)) ||
-                    (lastPart.isArray() && newPath.get(i).startsWith("["))) {
-                    pathMatches++;
-                } else {
-                    break;
-                }
-            }
-            return pathMatches;
-        }
-
-        private void close() throws IOException {
-            if ( !shallowObjects.isEmpty() ) {
-                for (String o : shallowObjects) {
-                    jg.writeFieldName(o);
-                    jg.writeBoolean(true);
-                }
-            }
-            while (jg.getOutputContext().getParent() != null) {
-                JsonStreamContext oc = jg.getOutputContext();
-                if (oc.inObject()) {
-                    jg.writeEndObject();
-                } else if (oc.inArray()) {
-                    jg.writeEndArray();
-                }
-            }
-            jg.flush();
-            if( options.callback()!=null ) {
-                output.write(")".getBytes(StandardCharsets.UTF_8));
-            }
-            jg.close();
-        }
-
-        private void writeValue(JsonRecord value) throws IOException {
-            switch (value.getKind()) {
-                case JsonTokenId.ID_STRING:
-                    jg.writeString(value.getValue());
-                    break;
-                case JsonTokenId.ID_NULL:
-                    jg.writeNull();
-                    break;
-                case JsonTokenId.ID_NUMBER_FLOAT:
-                case JsonTokenId.ID_NUMBER_INT:
-                    jg.writeNumber(value.getValue());
-                    break;
-                case JsonTokenId.ID_TRUE:
-                    jg.writeBoolean(true);
-                    break;
-                case JsonTokenId.ID_FALSE:
-                    jg.writeBoolean(false);
-                    break;
-                default:
-            }
-        }
-    }
 }
